Split frontend JSON smoke tests
