+++ tests/integration/cli_build/frontend_json/typed_json.rs
@@ -1,0 +1,52 @@
+use crate::support::*;
+use std::process::Command;
+
+#[test]
+fn emit_json_typed_command_outputs_checked_program() {
+    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
+        .args([
+            "emit-json",
+            "typed",
+            test_dir().join("generic_method.zen").to_str().unwrap(),
+        ])
+        .output()
+        .expect("run zen emit-json typed");
+
+    assert!(
+        output.status.success(),
+        "zen emit-json typed failed: stdout={}, stderr={}",
+        String::from_utf8_lossy(&output.stdout),
+        String::from_utf8_lossy(&output.stderr)
+    );
+
+    let json: serde_json::Value =
+        serde_json::from_slice(&output.stdout).expect("emit-json typed stdout is json");
+    assert_eq!(json["format"], "zen.typed.v0");
+    assert_eq!(json["semantic_status"], "checked");
+
+    let functions = json["program"]["functions"]
+        .as_array()
+        .expect("typed functions array");
+    assert!(
+        functions
+            .iter()
+            .any(|function| function["name"] == "Box.get_i32"),
+        "typed JSON should contain specialized generic method: {json}"
+    );
+    assert!(
+        functions.iter().any(|function| function["name"] == "main"),
+        "typed JSON should contain main function: {json}"
+    );
+
+    let types = json["program"]["types"]
+        .as_array()
+        .expect("typed types array");
+    assert!(
+        types.iter().any(|ty| ty["name"] == "Box_i32"),
+        "typed JSON should contain specialized generic type: {json}"
+    );
+
+    let serialized = String::from_utf8(output.stdout).expect("typed JSON is utf-8");
+    assert!(!serialized.contains("Box_T"));
+    assert!(!serialized.contains("T Box_get"));
+}
